llm-cot-siliconflow-py/requirements.txt
requests
pandas
matplotlib
seaborn
python-dotenv
notebook
tqdm